+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Prepare stub for Storage Indexer. Disable fetching packages from Package Storage v1. [#811](https://github.com/elastic/package-registry/pull/811)
-* Basic implement of storage indexer. [#814](https://github.com/elastic/package-registry/pull/814)
+* Basic implement of Storage Indexer. [#814](https://github.com/elastic/package-registry/pull/814)
 
 ### Deprecated